requirements.txt
absl-py==0.5.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.3
chardet==3.0.4
cheroot==6.5.2
CherryPy==18.0.1
cloudpickle==0.6.1
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
docx==0.2.4
feedparser==5.2.1
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
imutils==0.5.2
jaraco.functools==1.20
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lxml==4.2.5
Markdown==3.0.1
matplotlib==3.0.0
more-itertools==4.3.0
networkx==2.2
nlpnet==1.2.2
nltk==3.3
numpy==1.15.2
opencv-python==4.0.0.21
pandas==0.24.2
pattern3==3.0.0
pdfminer.six==20170720
pdfminer3k==1.3.1
Pillow==5.3.0
pluggy==0.8.0
ply==3.11
portend==2.3
progressbar2==3.38.0
protobuf==3.6.1
py==1.7.0
pycryptodome==3.7.0
pyparsing==2.2.2
pytest==3.9.3
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
simplejson==3.16.0
six==1.11.0
tempora==1.13
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
toolz==0.9.0
torch==1.1.0.post2
torchvision==0.3.0
Werkzeug==0.14.1
zc.lockfile==1.3.0
